ux: Relay Detail Redesign
This patch redesigns the relay detail view. The first step needed to further improve how relays are viewed.
In addition, Fees are added to the relay metadata to present to the user the admission, subscription, or publication fees a relay may have.

There are various changes made, but mainly several relay details are moved outside of the main RelayDetail view for easier tracking and updates.

With this design we will be able to add ratings and relay previews, as this patch is large enough I will submit that improvement in the future.

Testing
——
iPhone 15 Pro Max (17.3.1) Dark Mode:

https://i.nostr.build/VwVvq.png

https://i.nostr.build/KGO0v.png

iPhone SE (3rd generation) (16.4) Light Mode:

https://i.nostr.build/M5V26.png

https://i.nostr.build/gZKw3.png

——

Changelog-Added: Relay fees metadata
Changelog-Changed: Relay detail design

Signed-off-by: ericholguin <ericholguin@apache.org>